+++ packages/shared/components/src/Tag/index.ts
@@ -1,0 +1,4 @@
+/* eslint-disable @stylistic/padding-line-between-statements, react-refresh/only-export-components */
+export { TagIcon, type TagIconProps } from './tag-icon';
+export { TagLabel, type TagLabelProps } from './tag-label';
+export { Tag, type TagProps } from './tag';
